sparkzz/echoverse/requirements.txt
streamlit>=1.40.0
python-dotenv
requests
gtts
pillow
